+++ cli/src/coreutils/cd.ts
@@ -2,7 +2,6 @@
  * Discord Debate Timer
  *
  * @license BSD-3-Clause
- * @version 1.9.3
  * @author Luke Zhang luke-zhang-04.github.io/
  * @file lets You send messages on the bots behalf
  * @copyright 2020 - 2021 Luke Zhang