+++ src/commands/timer/index.ts
@@ -2,13 +2,12 @@
  * Discord Debate Timer
  *
  * @license BSD-3-Clause
- * @version 1.9.3
  * @author Luke Zhang luke-zhang-04.github.io/
  * @copyright 2020 - 2021 Luke Zhang
  */
 
+import * as dateplus from "@luke-zhang-04/dateplus/dist/cjs/dateplus.cjs"
 import {formatTime, muteUser} from "./utils"
-import DatePlus from "@luke-zhang-04/dateplus/dist/cjs/dateplus.cjs"
 import {verbosity} from "../../getConfig"
 
 /* eslint-disable no-use-before-define */
@@ -38,7 +37,7 @@ export class Timer {
     /**
      * If user should be muted after their speech temporarily (experimental)
      */
-    public shouldMute = true
+    public shouldMute: boolean
 
     /**
      * If timer is currently paused
@@ -117,18 +116,18 @@ export class Timer {
         timeCtrl: number,
         protectedTime?: number,
     ) {
-        this._mentionedUser = message.mentions.users.first() // Mentioned user
+        const mentionedMember = message.mentions.members?.first()
+
+        this._mentionedUser = mentionedMember?.user // Mentioned user
         this.mentionedUid = this._mentionedUser?.id // Id of aforementioned user
         this.creator = message.author
 
         const uid = this.mentionedUid
 
         // Make sure timer isn't longer than 15 mins
-        this.timeCtrl = isNaN(timeCtrl) ? DatePlus.minsToSecs(5) : timeCtrl
+        this.timeCtrl = isNaN(timeCtrl) ? dateplus.minsToSecs(5) : timeCtrl
 
-        console.log(protectedTime)
-
-        this.protectedTime = protectedTime ?? (this.timeCtrl >= DatePlus.minsToSecs(7) ? 60 : 30)
+        this.protectedTime = protectedTime ?? (this.timeCtrl >= dateplus.minsToSecs(7) ? 60 : 30)
 
         message.channel.send(`:timer: Starting timer${uid ? ` for debater <@${uid}>` : ""}!`)
 
@@ -138,7 +137,7 @@ export class Timer {
             process.env.NODE_ENV === "test" ? "" : `\`[${"\u2014".repeat(this._barWidth)}]\` 0%\n`
 
         const isProtectedTime =
-            this.timeCtrl <= DatePlus.minsToSecs(3) || (!this._stages[1] && !this._stages[4])
+            this.timeCtrl <= dateplus.minsToSecs(3) || (!this._stages[1] && !this._stages[4])
 
         this._msg = message.channel.send(
             `${bar}${timerTarget}Started by: ${this.creator}\nCurrent time: ${formatTime(
@@ -147,6 +146,14 @@ export class Timer {
                 isProtectedTime ? "yes" : "no"
             }\nId: ${this._fakeId ?? "unknown"}${this.isPaused ? "\nPaused" : ""}`,
         )
+
+        // A bunch of checks to make sure someone doesn't mute someone else on another server
+        this.shouldMute =
+            this._mentionedUser !== undefined &&
+            message.member !== null && // Make sure the message was sent from a server
+            message.guild !== null &&
+            message.member.permissions.has(["MUTE_MEMBERS"], true) && // Make sure this user has permissions to mute people
+            mentionedMember?.voice?.guild?.id === message.guild.id // Make sure the mentioned user is in a voice call in the same server
     }
 
     public get time(): number {
@@ -214,7 +221,7 @@ export class Timer {
             if (this.isPaused) {
                 this._startTime += Values.Interval * 1000
 
-                if (Date.now() - this._trueStartTime > DatePlus.minsToMs(20)) {
+                if (Date.now() - this._trueStartTime > dateplus.minsToMs(20)) {
                     this.message.channel.send(
                         `Timer with id ${this._fakeId} has been paused for more than 20 minutes. This timer is now being killed.`,
                     )
@@ -231,7 +238,7 @@ export class Timer {
             // If speech surpasses time
             if (
                 this.time >= this.timeCtrl + Values.GraceTime ||
-                this.time > DatePlus.minsToSecs(20)
+                this.time > dateplus.minsToSecs(20)
             ) {
                 this.kill(false)
             }
@@ -315,7 +322,7 @@ export class Timer {
                 : `\`[${blocks}${dashes}]\` ${percentage}%\n`
 
         const isProtectedTime =
-            this.timeCtrl <= DatePlus.minsToSecs(3) || (!this._stages[1] && !this._stages[4])
+            this.timeCtrl <= dateplus.minsToSecs(3) || (!this._stages[1] && !this._stages[4])
 
         msg.edit(
             // Edit the message with required information
@@ -338,7 +345,7 @@ export class Timer {
         const {time, timeCtrl} = this
 
         // 3 minute speechs are all protected
-        const hasProtectedTime = this.timeCtrl > DatePlus.minsToSecs(3)
+        const hasProtectedTime = this.timeCtrl > dateplus.minsToSecs(3)
 
         // The messages can be used as comments they're kinda self explanatory
         if (!this._stages[1] && hasProtectedTime && time >= this.protectedTime) {