+++ src/events/index.ts
@@ -1,11 +1,9 @@
-#!/bin/node
 /**
  * Discord Debate Timer
  *
  * @license BSD-3-Clause
- * @version 1.10.0
  * @author Luke Zhang luke-zhang-04.github.io/
  * @copyright 2020 - 2021 Luke Zhang
  */
 
-require("./lib")
+export {onVoiceStateUpdate} from "./onVoiceStateUpdate"