+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.0]
+
+### Added
+
+-   feat: add voice channel roles for prepping
+
+### Changed
+
+-   feat(config): use zod to validate config
+
+## [1.10.0]
+
+### Changed
+
+-   feat(cmd/timer): prevent muting members from other servers
+
+### Security
+
+-   fix(security): prevent ReDOS attacks with RE2
+
 ## [1.9.3] - 2021-04-13
 
 ### Fixed